Fix: Unable to set spell group to 'NONE'
Using the spellgroup("none"), didn't work as below in the code, after it was already loaded, the code: if (group == SPELLGROUP_NONE) {, would have overriden it and set none to either attack or healing group.

This successfully fixes it.

When group / subgroup is NONE, we dont need to load cooldowns for them, as they wont be used since they will be 0 then.